Added: comments to Pistol, Scope, TestPistol, and TestScope Updated: Tests in TestPistol

+++ lab4_Decorator/src/Weapon/Scope.java
@@ -6,9 +6,14 @@ public class Scope extends Attachment{
 		super(w);
 	}
 	
+	/**
+	 * Calculates the damage done only by the scope and returns as an int.
+	 * All values typecast to double so no accidental rounding occurs.
+	 * Final result is then typecast as an int.
+	 */
 	public int calculateDamage(int distance)
 	{
-		return (int) ((double)base.calculateDamage(distance) * ((double)(1 + (double)(base.range + 10) - distance)/(double)(base.range + 10)));
+		return (int) ((double)(base.calculateDamage(distance) * ((double)(1 +(base.range + 10) - distance)/(base.range + 10))));
 	}
 
 }
